# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth",
        "description": "Document about Authentification ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/users/create": {
            "post": {
                "tags": [
                    "create user"
                ],
                "summary": "Add user",
                "operationId": "d37ea9642c2cfbe867361ca7da44eec7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom",
                                    "prenom",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "hamza"
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "example": "saad"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "saad@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "Test@1245"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "users created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "users not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/update/{id}": {
            "put": {
                "tags": [
                    "update role of user"
                ],
                "summary": "Update user",
                "operationId": "97d6cee3e6d901306b2a691941b07ef0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role"
                                ],
                                "properties": {
                                    "role": {
                                        "type": "string",
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "user updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "users updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "user not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/delete/{id}": {
            "delete": {
                "tags": [
                    "delete user"
                ],
                "summary": "delete user",
                "operationId": "9715e31688901372e9b3959ff93d0422",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 101
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "user delete successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:20:00"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "users delete successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "user not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Failed"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "users not found"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "example": "2025-10-13 11:45:00"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}